requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-Limiter
argon2-cffi
stripe
python-dotenv
Flask-WTF
requests